+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.1.0
+* expose `isPointOnSegment` function
+* use `point-to-segment-2d` module
+* remove stale warning about node `< 12.x` which has been unsupported for a while now
+
+
 # 1.0.0
 * accidentally changed default behavior to allow segments <= 1 distance to be considered intersecting
 * refactor the API to provide a garbage collector friendly invocation and better performance (#2)